Extract common validations to a common module
