+++ packages/app/src/cli/models/extensions/specifications/app_config_app_proxy.ts
@@ -1,4 +1,4 @@
-import {validateUrl} from './validation/common.js'
+import {validateUrl} from '../../app/validation/common.js'
 import {ExtensionSpecification, createConfigExtensionSpecification} from '../specification.js'
 import {zod} from '@shopify/cli-kit/node/schema'
